requirements.txt
streamlit==1.28.2
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
pandas